Move 12 new shape functions to separate file
